Ensure Null integration can be loaded
