+++ lib/mutest.rb
@@ -179,6 +179,7 @@ require 'mutest/expression/methods'
 require 'mutest/expression/namespace'
 require 'mutest/test'
 require 'mutest/integration'
+require 'mutest/integration/null'
 require 'mutest/selector'
 require 'mutest/selector/expression'
 require 'mutest/config'
