+++ spec/unit/mutest/cli_spec.rb
@@ -137,7 +137,7 @@ Options:
     end
 
     context 'with use flag' do
-      context 'when integration exists' do
+      context 'when using the existing integration with rspec' do
         let(:flags) { %w[--use rspec] }
 
         before do
@@ -149,6 +149,19 @@ Options:
         it_should_behave_like 'a cli parser'
 
         let(:expected_integration) { Mutest::Integration::Rspec }
+      end
+
+      context 'when specifying the default null integration explicitely' do
+        let(:flags) { %w[--use null] }
+        let(:expected_integration) { Mutest::Integration::Null }
+
+        before do
+          expect(Kernel).to receive(:require)
+            .with('mutest/integration/null')
+            .and_call_original
+        end
+
+        it_should_behave_like 'a cli parser'
       end
 
       context 'when integration does NOT exist' do
